Make position hash nonzero
Made PositionHash nonzero.
This will allow memory optimizations in transposition table.
